+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 -------------------
 
 - Properly handle long path names on Windows when downloading a scenario. (nens/rana#5008)
+- Extend dialog for importing schematisation from HCC with revision picker. (nens/rana#5035)
 
 
 1.2.20 (2026-09-23)